+++ src/FastSimulation.java
@@ -1,0 +1,14 @@
+public class FastSimulation extends Symulacja {
+    private final Human[] humans;
+    private final Animal[] animals;
+
+
+    public FastSimulation(Human[] humans, Animal[] animals) {
+        this.humans = humans;
+        this.animals = animals;
+    }
+    public void fastsimulation(){
+
+    }
+}
+
